pyproject.toml
[project]
name = "job-template-run"
version = "0.17.0"
description = "A Kubernetes controller for templated jobs and creating job runs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["httpx>=0.28.0", "kopf>=1.37.5", "kubernetes>=32.0.1", "pyyaml>=6.0.2"]

[dependency-groups]
dev = [
    "pyright>=1.1.399",
    "pytest>=8.0.0",
    "ruff>=0.11.5",
    "kubernetes-stubs-elephant-fork>=32.0.1",
    "commitizen>=4.6.0",
    "cz-changeup>=1.1.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.commitizen]
name = "cz_changeup"
version_provider = "pep621"
version_files = [
    "charts/job-template-run/Chart.yaml",
    "charts/job-template-run/values.yaml",
]
pre_bump_hooks = [
    'sed -i "/name = \"job-template-run\"/{n;s/version = \".*\"/version = \"$CZ_PRE_NEW_VERSION\"/}" uv.lock',
]
ignored_tag_formats = ["job-template-run-*"]

changeup_repo_base_url = "https://github.com/hematoscope/job-template-run"
